[IMP] web: support many2many grouping
PURPOSE

Allow to group records by many2many fields.

Use case: turn the user_id m2o field on project.task into a user_ids m2m field
in order to assign multiple collaborators on a task

SPECIFICATION

Add many2many fields to the 'list of groupable fields'.
When grouping a set of records by a many2many, records linked to multiple
records of the comodel will essentially be in multiple 'groups' (even though
those are not really 'groups' anymore).
There will also be a "null" group, for records that have an empty value of the
many2many, like "unassigned" records.

Example: grouping a project.task linked to 10 res.users by the user_ids
many2many field, the project.task will be 'rendered/considered/displayed' 10
times, once per res.users.
Kind of like if we grouped the m2m relation table instead.

When grouping by a m2m field:
- listview
    editing an instance of a record also updates each other instance(if any)
        - the same goes when scheduling an activity, etc.
    disable deletion
        - the risk here is that the user might expected that this will 'unlike'
        the record from a specific record of the many2many field
    disable archive/unarchive
- kanban
    disable drag&drop between groups
    disable deletion
    disable archive/unarchive
- searchview
    display many2many fields custom groupby dropdown

With this commit, we also support groupby many2many field in mockServer so that
test cases are written based on that.

task-2508608

+++ addons/web/static/src/legacy/js/views/list/list_controller.js
@@ -423,6 +423,13 @@ var ListController = BasicController.extend({
      * @private
      */
     _getActionMenuItems: function (state) {
+        let isM2MGrouped = false;
+        if (state.groupedBy.length) {
+            isM2MGrouped = state.groupedBy.some((group) => {
+                const groupByFieldName = group.split(':')[0];
+                return state.fields[groupByFieldName].type === "many2many";
+            });
+        }
         if (!this.hasActionMenus || !this.selectedRecords.length) {
             return null;
         }
@@ -434,7 +441,7 @@ var ListController = BasicController.extend({
                 callback: () => this._onExportData()
             });
         }
-        if (this.archiveEnabled) {
+        if (this.archiveEnabled && !isM2MGrouped) {
             otherActionItems.push({
                 description: _t("Archive"),
                 callback: () => {
@@ -450,7 +457,7 @@ var ListController = BasicController.extend({
                 callback: () => this._toggleArchiveState(false)
             });
         }
-        if (this.activeActions.delete) {
+        if (this.activeActions.delete && !isM2MGrouped) {
             otherActionItems.push({
                 description: _t("Delete"),
                 callback: () => this._onDeleteSelectedRecords()
@@ -552,13 +559,22 @@ var ListController = BasicController.extend({
      * @override
      * @param {string} recordId
      */
-    _saveRecord: function (recordId) {
+    _saveRecord: function (recordId, options) {
+        options = options || {};
         var record = this.model.get(recordId, { raw: true });
         if (record.isDirty() && this.renderer.isInMultipleRecordEdition(recordId)) {
             // do not save the record (see _saveMultipleRecords)
             const prom = this.multipleRecordsSavingPromise || Promise.reject();
             this.multipleRecordsSavingPromise = null;
             return prom;
+        }
+        const state = this.model.get(this.handle, { raw: true });
+        if (state.groupedBy.length) {
+            const isM2MGrouped = state.groupedBy.some((group) => {
+                const groupByFieldName = group.split(':')[0];
+                return state.fields[groupByFieldName].type === "many2many";
+            });
+            options.isM2mGrouped = isM2MGrouped;
         }
         return this._super.apply(this, arguments);
     },
@@ -918,7 +934,27 @@ var ListController = BasicController.extend({
      */
     _onSaveLine: function (ev) {
         this.saveRecord(ev.data.recordID)
-            .then(ev.data.onSuccess)
+            .then((result) => {
+                const state = this.model.get(this.handle, { raw: true });
+                let isM2MGrouped = false;
+                if (state.groupedBy.length) {
+                    isM2MGrouped = state.groupedBy.some((group) => {
+                        const groupByFieldName = group.split(':')[0];
+                        return state.fields[groupByFieldName].type === "many2many";
+                    });
+                }
+                if (isM2MGrouped) {
+                    return this.update({}, { reload: false }).then(() => {
+                        if (ev.data.onSuccess) {
+                            ev.data.onSuccess();
+                        }
+                    });
+                } else {
+                    if (ev.data.onSuccess) {
+                        ev.data.onSuccess(result);
+                    }
+                }
+            })
             .guardedCatch(ev.data.onFailure);
     },
     /**
